Split up completers into separate files
